Update to 1.4.21
Mock was last updated for Clear Linux OS in Nov 2015, so it is overdue
for update. This commit bumps it to the latest release, 1.4.21.

The patch series adds a toplevel Makefile (adapted from the upstream
Fedora spec file) to enable the package for autospec and ports the
stateless integration changes from the previous patch series.

This mock release and the several previous releases have been tested
with local builds and with koji 1.17.1.

In addition to numerous bug fixes, here are some changes to highlight:

- `systemd-nspawn` is used by default now. To force it to use chroot,
  pass `--old-chroot` or set the "use_nspawn" config option to False.

- Config options can now be set on the command line with
  --config-opts=KEY=VALUE. To set multiple options, pass the option
  multiple times with different KEYs and VALUEs. These options will
  override settings from configuration files.

- `mockchain` has been deprecated. Use `mock --chain` instead.

- New log files are generated in the results directory: hw_info.log and
  installed_pkgs.log. The first log displays a summary of CPU,
  memory, and storage info of the host system performing a build. The
  second log lists all packages installed in the build root as build
  dependencies; previously, one would need to carefully scrape root.log
  to capture this info.

- Jinja2 templating is supported in config files.

- A bootstrap chroot can now be configured for builds. This feature will
  not be enabled in Clear Linux OS for now, but may be useful in the
  future to enable a more controlled build environment.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>